chore: code review TODO + refactor worker into modular packages
- Add comprehensive TODO.md with 37 findings from code review
  (4 critical, 11 high, 12 medium, 10 low across security, reliability, UX)
- Refactor monolithic file_processor.py into worker/file/ package
  (loader, chunker, processor, cli modules)
- Refactor monolithic web_processor.py into worker/web/ package
  (crawler, processor, cli modules)
- Add worker CLI entrypoint and deblurrer module
- Remove deprecated video_ingest package
