+++ worker/env.template
@@ -30,6 +30,12 @@ VLM_API_BASE=http://host.docker.internal:8000/v1
 VLM_API_KEY=sk-dummy-key
 VLM_MODEL=llava-v1.6-34b
 
+# ----- Deblurring (optional, before VLM/CV) -----
+DEBLUR_API_BASE=
+DEBLUR_API_KEY=
+DEBLUR_MODEL=deblur-gan-inception
+DEBLUR_NAME=none
+
 # ----- Worker Configuration -----
 WORKER_CONCURRENCY=5
 JOB_TIMEOUT=600
