--- worker/video_ingest/__init__.py
@@ -1,1 +1,0 @@
-# Launcher for: python -m video_ingest <video.mp4> --mode procedure --out chunks.jsonl
