--- worker/video_ingest/__main__.py
@@ -1,5 +1,0 @@
-"""Run: python -m video_ingest <video.mp4> --mode procedure --out chunks.jsonl"""
-from worker.video.cli import main
-
-if __name__ == "__main__":
-    raise SystemExit(main())
